Commit of statistics updated

+++ app/src/main/java/com/example/pazzly/domain/entity/MojBroj.java
@@ -5,7 +5,7 @@ public class MojBroj extends Game {
     public MojBroj() {
     }
 
-    public MojBroj(String gameName, int rounds, int maxPointsPerRound, int minimalPointsPerRound, int durationPerRound) {
+    public MojBroj(String gameName, int rounds, int maxPointsPerRound, int minimalPointsPerRound, double durationPerRound) {
         super(gameName, rounds, maxPointsPerRound, minimalPointsPerRound, durationPerRound, 1);
     }
 }
